Move utils up
