+++ packages/heuristics/src/detectors/cra.js
@@ -1,4 +1,4 @@
-const { hasRequiredDeps, hasRequiredFiles, packageManagerCommand, scanScripts } = require('./utils/jsdetect')
+const { hasRequiredDeps, hasRequiredFiles, packageManagerCommand, scanScripts } = require('../utils/jsdetect')
 
 /**
  * detection logic - artificial intelligence!
