+++ packages/heuristics/src/detectors/quasar-v0.17.js
@@ -1,4 +1,4 @@
-const { hasRequiredDeps, hasRequiredFiles, packageManagerCommand, scanScripts } = require('./utils/jsdetect')
+const { hasRequiredDeps, hasRequiredFiles, packageManagerCommand, scanScripts } = require('../utils/jsdetect')
 
 module.exports = function() {
   // REQUIRED FILES
